move config parser to module
